US1.4 - L'ajout en base de donnée de l'élève/tuteur/l'inscription

+++ src/main/java/geschool/persistence/interfaces/EleveDAO.java
@@ -21,5 +21,5 @@ public interface EleveDAO {
     Eleve rechercherUnEleveAvecId(Integer idEleve);
     Eleve rechercherUnEleveAvecMatricule(String matricule);
     Eleve rechercherUnEleveAvecNomEtPrenom(String nom,String prenom);
-    Integer rechercherLeNombreTotalEleve();
+    Long rechercherLeNombreTotalEleve();
 }
